+++ CHANGELOG.md
@@ -5,15 +5,18 @@
 * Replace the `YBSlantedCollectionViewLayout` with `CollectionViewSlantedLayout`
 * Replace the `YBSlantedCollectionViewCell` with `CollectionViewSlantedCell`
 
-
 #### Added
 * Add the `slantingDirection` property
 * Add the `itemSize` property
 * Add the `slantingAngle` property
 * Introduce the `CollectionViewDelegateSlantedLayout` delegate.
+* Add the `isFistCellExcluded` instead of `firstCellSlantingEnabled`
+* Add the `isLastCellExcluded` instead of `lastCellSlantingEnabled`
 
 #### removed
 * Remove the `reverseSlantingAngle` property
+* Remove the `firstCellSlantingEnabled` property
+* Remove the `lastCellSlantingEnabled` property
 * Remove the `itemSizeOptions` property and the `YBSlantedCollectionViewLayoutSizeOptions` struct.
 
 ## 2.2.0